Thread the covariate through both seams and the Rust backend (#658 items 2, 2b)
**The two Python seams.** `external_field` and `marginal_log_likelihood_torch`
scored without a covariate. `external_field` is a defect rather than deferred
work: `fit_spatio_sequential` passes it a posterior computed *with* the
covariate while its own `log_density` scored *without*, so the two halves of
one step disagreed about the model -- monotone either way, which is why nothing
failed. `covariate_block` is the one slice all three seams share; it adds the
trailing singleton only where the covariate carries no axes of its own, since a
`(S, V, 2)` covariate made `(S, V, 2, 1)` broadcasts against the wrong axis.

**The Rust backend, built out rather than refused.** The kernel tabulated
`log p(count | class, state)` and indexed by the count. Under a covariate the
density is a function of the count *and* what it is scored against, so no table
indexed by the count alone exists -- which is what made this look like a kernel
redesign. It is not. The kernel never knew what the index meant, only that the
table was built from it, so the caller tabulates the distinct `(count,
covariate)` pairs and hands the row each observation falls in.

`src/coupled.rs` changes in one way: the index widens from `u16` to `u32`,
because a row is no longer bounded by the largest count. `usize::from` is not
defined for `u32`, so `row_index` states the widening once rather than as a
cast at four sites. Without a covariate the rows *are* the counts and the table
is count-major -- the same numbers as before, asserted on the arrays and not
only on a result.

At the ci instance the tabulation finds 9,989 and 245 rows against 64,000
observations, so a per-vertex exposure repeated down the positions stays a
table rather than becoming a per-observation score array. Against the NumPy
oracle under a covariate: 2.7e-15 relative on the evidence, 2.9e-15 on the
field, 1.0e-09 absolute on the posteriors, one thread. Uncovaried: 3.109e-15,
the figure this module already carried.

`SpatioSequentialParams` validates the covariate's leading two axes only. What
follows them is the family's, as it is for the observations: none for a scalar
family, a channel axis for the count pair.

Two tests that had gone stale with the widening and with #656: one built
`uint16` arrays for the binding, and one matched a refusal message #656 split
in two. The second is red on `main` and carries its own pull request.

ruff, ruff format, mypy --strict, clippy, cargo fmt clean; cargo test 38
passed; 338 critical in 58.49 s; the rust, fit and kernel modules 24 passed.

+++ benches/oxi_snakes_and_ladders_bench.rs
@@ -259,11 +259,11 @@ fn bench_class_posteriors(c: &mut Criterion) {
         total: &total,
         success: &success,
     };
-    let totals: Vec<u16> = (0..n_positions * n_nodes)
-        .map(|index| (index % extent) as u16)
+    let totals: Vec<u32> = (0..n_positions * n_nodes)
+        .map(|index| (index % extent) as u32)
         .collect();
-    let successes: Vec<u16> = (0..n_positions * n_nodes)
-        .map(|index| ((index * 7) % extent) as u16)
+    let successes: Vec<u32> = (0..n_positions * n_nodes)
+        .map(|index| ((index * 7) % extent) as u32)
         .collect();
     let labels: Vec<i64> = (0..n_nodes).map(|v| (v % n_classes) as i64).collect();
     let log_initial = vec![-(n_states as f64).ln(); n_classes * n_states];
